Break up the eos_unit_test file into multiple files for ease of organization etc
