+++ test/test_eos_unit.cpp
@@ -27,32 +27,12 @@
 #include <singularity-eos/eos/eos.hpp>
 #include <singularity-eos/eos/eos_builder.hpp>
 
-// typename demangler
-#ifdef __GNUG__
-#include <cstdlib>
-#include <cxxabi.h>
-#include <memory>
-
-std::string demangle(const char *name) {
-
-  int status = -4; // some arbitrary value to eliminate the compiler warning
-
-  // enable c++11 by passing the flag -std=c++11 to g++
-  std::unique_ptr<char, void (*)(void *)> res{
-      abi::__cxa_demangle(name, NULL, NULL, &status), std::free};
-
-  return (status == 0) ? res.get() : name;
-}
-
-#else
-
-// does nothing if not g++
-std::string demangle(const char *name) { return name; }
-
+#ifndef CATCH_CONFIG_RUNNER
+#include "catch2/catch.hpp"
 #endif
 
-#define CATCH_CONFIG_RUNNER
-#include "catch2/catch.hpp"
+
+#include <test/eos_unit_test_helpers.hpp>
 
 using singularity::EOS;
 using singularity::Gruneisen;
@@ -85,56 +65,6 @@ constexpr int gID = 2700;
 constexpr Real ev2k = 1.160451812e4;
 #endif // SINGULARITY_TEST_SESAME
 #endif // SPINER_USE_HDF
-
-PORTABLE_INLINE_FUNCTION bool isClose(Real a, Real b, Real eps = 5e-2) {
-  return fabs(b - a) / (fabs(a + b) + 1e-20) <= eps;
-}
-
-PORTABLE_INLINE_FUNCTION Real myAtan(Real x, Real shift, Real scale, Real offset) {
-  return scale * atan(x - shift) + offset;
-}
-
-// Function for comparing arrays and outputting the important information
-template <typename X, typename Y, typename Z, typename ZT, typename XN, typename YN>
-inline void array_compare(int num, X &&x, Y &&y, Z &&z, ZT &&ztrue, XN xname, YN yname,
-                          Real tol = 1e-12) {
-  for (int i = 0; i < num; i++) {
-    INFO("i: " << i << ", " << xname << ": " << x[i] << ", " << yname << ": " << y[i]
-               << ", Value: " << z[i] << ", True Value: " << ztrue[i]);
-    CHECK(isClose(z[i], ztrue[i], 1e-12));
-  }
-}
-
-inline void compare_two_eoss(const EOS &test_e, const EOS &ref_e) {
-  // compare all individual member functions with 1 as inputs,
-  // this function is meant to catch mis-implementations of
-  // modifiers that can be initialized in such a way as to
-  // be equivalent of an unmodified eos. Best used with analytic
-  // eoss.
-  REQUIRE(isClose(test_e.TemperatureFromDensityInternalEnergy(1, 1),
-                  ref_e.TemperatureFromDensityInternalEnergy(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.InternalEnergyFromDensityTemperature(1, 1),
-                  ref_e.InternalEnergyFromDensityTemperature(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.PressureFromDensityInternalEnergy(1, 1),
-                  ref_e.PressureFromDensityInternalEnergy(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.SpecificHeatFromDensityInternalEnergy(1, 1),
-                  ref_e.SpecificHeatFromDensityInternalEnergy(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.BulkModulusFromDensityInternalEnergy(1, 1),
-                  ref_e.BulkModulusFromDensityInternalEnergy(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.GruneisenParamFromDensityInternalEnergy(1, 1),
-                  ref_e.GruneisenParamFromDensityInternalEnergy(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.PressureFromDensityTemperature(1, 1),
-                  ref_e.PressureFromDensityTemperature(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.SpecificHeatFromDensityTemperature(1, 1),
-                  ref_e.SpecificHeatFromDensityTemperature(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.BulkModulusFromDensityTemperature(1, 1),
-                  ref_e.BulkModulusFromDensityTemperature(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.GruneisenParamFromDensityTemperature(1, 1),
-                  ref_e.GruneisenParamFromDensityTemperature(1, 1), 1.e-15));
-  REQUIRE(isClose(test_e.MinimumDensity(), ref_e.MinimumDensity(), 1.e-15));
-  REQUIRE(isClose(test_e.MinimumTemperature(), ref_e.MinimumTemperature(), 1.e-15));
-  return;
-}
 
 SCENARIO("Test that fast logs are invertible and run on device", "[FastMath]") {
   GIVEN("A set of values to invert over a large dynamic range") {
@@ -633,225 +563,6 @@ SCENARIO("Vector EOS", "[VectorEOS][IdealGas]") {
   }
 }
 
-SCENARIO("Gruneisen EOS", "[VectorEOS][GruneisenEOS]") {
-  GIVEN("Parameters for a Gruneisen EOS") {
-    // Unit conversions
-    constexpr Real cm = 1.;
-    constexpr Real us = 1e-06;
-    constexpr Real Mbcc_per_g = 1e12;
-    // Gruneisen parameters for copper
-    constexpr Real C0 = 0.394 * cm / us;
-    constexpr Real S1 = 1.489;
-    constexpr Real S2 = 0.;
-    constexpr Real S3 = 0.;
-    constexpr Real Gamma0 = 2.02;
-    constexpr Real b = 0.47;
-    constexpr Real rho0 = 8.93;
-    constexpr Real T0 = 298.;
-    constexpr Real P0 = 0.;
-    constexpr Real Cv = 0.383e-05 * Mbcc_per_g;
-    // Create the EOS
-    EOS host_eos = Gruneisen(C0, S1, S2, S3, Gamma0, b, rho0, T0, P0, Cv);
-    EOS eos = host_eos.GetOnDevice();
-    GIVEN("Densities and energies") {
-      constexpr int num = 4;
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-      // Create Kokkos views on device for the input arrays
-      Kokkos::View<Real[num]> v_density("density");
-      Kokkos::View<Real[num]> v_energy("density");
-#else
-      // Otherwise just create arrays to contain values and create pointers to
-      // be passed to the functions in place of the Kokkos views
-      std::array<Real, num> density;
-      std::array<Real, num> energy;
-      auto v_density = density.data();
-      auto v_energy = energy.data();
-#endif // PORTABILITY_STRATEGY_KOKKOS
-
-      // Populate the input arrays
-      portableFor(
-          "Initialize density and energy", 0, 1, PORTABLE_LAMBDA(int i) {
-            v_density[0] = 8.0;
-            v_density[1] = 9.0;
-            v_density[2] = 9.5;
-            v_density[3] = 0.;
-            v_energy[0] = 1.e9;
-            v_energy[1] = 5.e8;
-            v_energy[2] = 1.e8;
-            v_energy[3] = 0.;
-          });
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-      // Create host-side mirrors of the inputs and copy the inputs. These are
-      // just used for the comparisons
-      auto density = Kokkos::create_mirror_view(v_density);
-      auto energy = Kokkos::create_mirror_view(v_energy);
-      Kokkos::deep_copy(density, v_density);
-      Kokkos::deep_copy(energy, v_energy);
-#endif // PORTABILITY_STRATEGY_KOKKOS
-
-      // Gold standard values for a subset of lookups
-      constexpr std::array<Real, num> pressure_true{
-          -1.282094800000000e+11, 1.998504088912181e+10, 9.595823319513451e+10,
-          P0 - C0 * C0 * rho0};
-      constexpr std::array<Real, num> bulkmodulus_true{
-          9.990648504000005e+11, 1.460692677162573e+12, 1.851227213843747e+12,
-          Gamma0 * (P0 - C0 * C0 * rho0)};
-      constexpr std::array<Real, num> temperature_true{
-          5.590966057441253e+02, 4.285483028720627e+02, 3.241096605744125e+02, T0};
-      constexpr std::array<Real, num> gruneisen_true{Gamma0, 2.007944444444444e+00,
-                                                     1.927000000000000e+00, Gamma0};
-
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-      // Create device views for outputs and mirror those views on the host
-      Kokkos::View<Real[num]> v_temperature("Temperature");
-      Kokkos::View<Real[num]> v_pressure("Pressure");
-      Kokkos::View<Real[num]> v_bulkmodulus("bmod");
-      Kokkos::View<Real[num]> v_gruneisen("Gamma");
-      auto h_temperature = Kokkos::create_mirror_view(v_temperature);
-      auto h_pressure = Kokkos::create_mirror_view(v_pressure);
-      auto h_bulkmodulus = Kokkos::create_mirror_view(v_bulkmodulus);
-      auto h_gruneisen = Kokkos::create_mirror_view(v_gruneisen);
-#else
-      // Create arrays for the outputs and then pointers to those arrays that
-      // will be passed to the functions in place of the Kokkos views
-      std::array<Real, num> h_temperature;
-      std::array<Real, num> h_pressure;
-      std::array<Real, num> h_bulkmodulus;
-      std::array<Real, num> h_gruneisen;
-      // Just alias the existing pointers
-      auto v_temperature = h_temperature.data();
-      auto v_pressure = h_pressure.data();
-      auto v_bulkmodulus = h_bulkmodulus.data();
-      auto v_gruneisen = h_gruneisen.data();
-#endif // PORTABILITY_STRATEGY_KOKKOS
-
-      WHEN("A P(rho, e) lookup is performed") {
-        eos.PressureFromDensityInternalEnergy(v_density, v_energy, v_pressure, num);
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-        Kokkos::fence();
-        Kokkos::deep_copy(h_pressure, v_pressure);
-#endif // PORTABILITY_STRATEGY_KOKKOS
-        THEN("The returned P(rho, e) should be equal to the true value") {
-          array_compare(num, density, energy, h_pressure, pressure_true, "Density",
-                        "Energy");
-        }
-      }
-
-      WHEN("A B_S(rho, e) lookup is performed") {
-        eos.BulkModulusFromDensityInternalEnergy(v_density, v_energy, v_bulkmodulus, num);
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-        Kokkos::fence();
-        Kokkos::deep_copy(h_bulkmodulus, v_bulkmodulus);
-#endif // PORTABILITY_STRATEGY_KOKKOS
-        THEN("The returned B_S(rho, e) should be equal to the true value") {
-          array_compare(num, density, energy, h_bulkmodulus, bulkmodulus_true, "Density",
-                        "Energy");
-        }
-      }
-
-      WHEN("A T(rho, e) lookup is performed") {
-        eos.TemperatureFromDensityInternalEnergy(v_density, v_energy, v_temperature, num);
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-        Kokkos::fence();
-        Kokkos::deep_copy(h_temperature, v_temperature);
-#endif // PORTABILITY_STRATEGY_KOKKOS
-        THEN("The returned B_S(rho, e) should be equal to the true value") {
-          array_compare(num, density, energy, h_temperature, temperature_true, "Density",
-                        "Energy");
-        }
-      }
-
-      WHEN("A Gamma(rho, e) lookup is performed") {
-        eos.GruneisenParamFromDensityInternalEnergy(v_density, v_energy, v_gruneisen,
-                                                    num);
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-        Kokkos::fence();
-        Kokkos::deep_copy(h_gruneisen, v_gruneisen);
-#endif // PORTABILITY_STRATEGY_KOKKOS
-        THEN("The returned Gamma(rho, e) should be equal to the true value") {
-          array_compare(num, density, energy, h_gruneisen, gruneisen_true, "Density",
-                        "Energy");
-        }
-      }
-    }
-  }
-}
-
-SCENARIO("Aluminum Gruneisen EOS Sound Speed Comparison", "[GruneisenEOS]") {
-  GIVEN("Parameters for a Gruneisen EOS") {
-    // Unit conversions
-    constexpr Real mm = 10.;
-    constexpr Real cm = 1.;
-    constexpr Real us = 1.e-06;
-    constexpr Real Mbar = 1.e12;
-    constexpr Real Mbcc_per_g = 1e12;
-    // Gruneisen parameters for copper
-    constexpr Real C0 = 0.535 * cm / us;
-    constexpr Real S1 = 1.34;
-    constexpr Real S2 = 0.;
-    constexpr Real S3 = 0.;
-    constexpr Real Gamma0 = 1.97;
-    constexpr Real b = 0.;
-    constexpr Real rho0 = 2.714000;
-    constexpr Real T0 = 298.;
-    constexpr Real P0 = 1e-06 * Mbar;
-    constexpr Real Cv = 0.383e-05 * Mbcc_per_g;
-    // Create the EOS
-    EOS host_eos = Gruneisen(C0, S1, S2, S3, Gamma0, b, rho0, T0, P0, Cv);
-    EOS eos = host_eos.GetOnDevice();
-    GIVEN("Density and energy") {
-      constexpr Real density = 5.92418956756592;            // g/cm^3
-      constexpr Real energy = 792486007.804619;             // erg/g
-      constexpr Real true_pres = 2.620656373250729;         // Mbar
-      constexpr Real true_sound_speed = 1.5247992468363685; // cm/us
-      WHEN("A P(rho, e) lookup is performed") {
-        Real pres = eos.PressureFromDensityInternalEnergy(density, energy);
-        THEN("The correct pressure should be returned") {
-          pres = pres / Mbar;
-          INFO("Density: " << density << "  Energy: " << energy << "  Pressure: " << pres
-                           << " Mbar  True pressure: " << true_pres << " Mbar");
-          REQUIRE(isClose(pres, true_pres, 1e-12));
-        }
-      }
-      WHEN("A B_S(rho, e) lookup is performed") {
-        const Real bulk_modulus =
-            eos.BulkModulusFromDensityInternalEnergy(density, energy);
-        THEN("The correct sound speed should be computed") {
-          const Real sound_speed = std::sqrt(bulk_modulus / density) / (cm / us);
-          INFO("Density: " << density << "  Energy: " << energy
-                           << "  Sound speed: " << sound_speed
-                           << " cm/us  True sound speed: " << true_sound_speed
-                           << " cm/us");
-          REQUIRE(isClose(sound_speed, true_sound_speed, 1e-12));
-        }
-      }
-      WHEN("A finite difference approximation is used for the bulk modulus") {
-        // Bulk modulus approximation:
-        //  B_S = rho * dPdr_e + P / rho * dPde_r
-        constexpr Real drho = 1e-06 * density;
-        constexpr Real de = 1e-06 * energy;
-        const Real P1 = eos.PressureFromDensityInternalEnergy(density, energy);
-        Real P2 = eos.PressureFromDensityInternalEnergy(density + drho, energy);
-        const Real dPdr_e = (P2 - P1) / drho;
-        P2 = eos.PressureFromDensityInternalEnergy(density, energy + de);
-        const Real dPde_r = (P2 - P1) / de;
-        const Real bmod_approx = density * dPdr_e + P1 / density * dPde_r;
-        THEN("The finite difference solution should approximate the exact solution") {
-          const Real bulk_modulus =
-              eos.BulkModulusFromDensityInternalEnergy(density, energy);
-          const Real ss_approx = std::sqrt(bmod_approx / density);
-          const Real sound_speed = std::sqrt(bulk_modulus / density);
-          INFO("Density: " << density << "  Energy: " << energy
-                           << "  Sound speed: " << sound_speed
-                           << " cm/us  Approximate sound speed: " << ss_approx
-                           << " cm/us");
-          REQUIRE(isClose(sound_speed, ss_approx, 1e-5));
-        }
-      }
-    }
-  }
-}
-
 #ifdef SPINER_USE_HDF
 #ifdef SINGULARITY_TEST_SESAME
 #ifdef SINGULARITY_USE_EOSPAC
@@ -1280,16 +991,3 @@ SCENARIO("Stellar Collapse EOS", "[StellarCollapse][EOSBuilder]") {
 }
 #endif // SINGULARITY_TEST_STELLAR_COLLAPSE
 #endif // USE_HDF5
-
-int main(int argc, char *argv[]) {
-
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-  Kokkos::initialize();
-#endif
-  int result;
-  { result = Catch::Session().run(argc, argv); }
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-  Kokkos::finalize();
-#endif
-  return result;
-}
